QR CODE SCANNER/requirements.txt
qrcode>=7.4
Pillow>=10.0
pyzbar>=0.1.9
opencv-python>=4.8  # optional fallback decoder on all platforms


